De-duplicate issues before first write to taskwarrior.
Fixes #601.

This can happen if, for instance, you have two github targets declared which
query in different ways.  They may each find the same issue which (without this
patch) may both be written to taskwarrior as duplicates.  After duplicates are
in place, it is difficult for bugwarrior to figure out which one to apply
annotation updates to.